Revert "[RISC-V] Add misaligned load store tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/512a5dd0-7d66-4d86-8a35-22a1a35e90ce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I26a2ce479e3b31fd25200070d1650749a3ff5ce4