Revert "[RISC-V] Add misaligned load store tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0af2332b-af5b-43a9-8145-bae8e9b00bf4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieb6334d5bc7914bd4487c20e3461501c31f4c5fd