Revert "[RISC-V] Add misaligned load store tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8ace34cb-9a3d-4772-8833-5e49f451382e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1a1bcced5fb9eeaa599d4dbac961d54472c7e35f