Revert "Switch the loader to a noexcept version of libc++"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f3360644-d24b-437a-b673-3f7f69a1b776). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9a4463d59cd312880134cc009f12f33f85b2ef68