Revert "Switch the loader to a noexcept version of libc++"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/616536ed-9fe7-447a-b229-f101f8e421c1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifc5e9f33b9bcaef07197fec8d8ed2d68b76e0d34