Revert "Switch the loader to a noexcept version of libc++"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c323cbd-2178-4052-8796-e826cca76e34). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaeb00f66eb2ccc649bb7b6ffc33b3b626c36980f