Revert "Switch the loader to a noexcept version of libc++"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cf17be33-4739-4891-a52d-5180d4fcf97c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7abeab3c5a9e826efc34dd4d18c91c46c3291a35