Revert "Switch the loader to a noexcept version of libc++"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/57490344-2420-4f34-975d-8310c0d41ee3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I24ddc9f1647b38ca36ca6855948a8ed7b3f6fbf4