Revert "Switch the loader to a noexcept version of libc++"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7b43dcdf-0dd9-4391-bdf0-e1e764c7c00c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ice0c6b782974766aec651bdf806fa77aa2c209fe