Revert "[RISC-V] Add misaligned load store tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ec8bf3fa-368c-49fa-a26d-c61078f599fa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idd66aca3639f9a2c79e385fe159a5328cc2ceded