Revert "[RISC-V] Add misaligned load store tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aeeacd72-d01c-41ce-8301-c867ba977414). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4f69a32aa66f66b632d9f8580463752a8eecb21e