Revert "Switch the loader to a noexcept version of libc++"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3563e708-b85f-41d8-bf93-3f4c1705a97a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I233f69bfff187e99f22e449e127417d3ba18778f