<sys/thread_properties.h>: fix tests.
These tests weren't being compiled, but they also didn't compile (in
particular, `__BIONIC__` is defined by <features.h>, not the toolchain),
and didn't work.

The __libc_register_thread_exit_callback() test was broken because it
specifically doesn't work for the main thread, but was only testing the
main thread. I've also updated the docs to clarify this. I trust the
implementation over the test because hwasan has similar hard-coded stuff
that also ignores the main thread.

Test: treehugger
Change-Id: I202f42ec0f623ee39b2922d921ac07261994072c