Revert "Convert Android.mk under bionic to Android.bp"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/627b9854-a1d6-497c-b340-1085a4336cbd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 342031208
Change-Id: I851026026d1731442b4303269445acb33566e768
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com