Revert "Convert Android.mk under bionic to Android.bp"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4af82619-85c8-4ba6-8d92-9ffec590c1fe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 342029467
Change-Id: I635ec6bff0a49c1fef2200a7e989b9bd7bf54241
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com