Revert "<unistd.h>: explain why getpagesize() is better than get..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/12c86bf0-8d25-417c-b182-88e177da9b50). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id51d74f2675534d2148a14efd13ff54e83b06d8d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com