[scudo] Implementation of Scudo backend for cache data retrieval. Cache capacity now defaults to maximum cache capacity when user configuration sets cache capacity to be greater than maximum cache capacity.
Initially, the Scudo allocator would return an error if the user attempted to set the cache capacity (i.e. the number of possible entries in the cache) above the maximum cache capacity. Now the allocator will resort to using the maximum cache capacity in this event. An interface has been set up to allow the user to query the maximum cache capacity as well as the currently set capacity.

Test: adb shell /data/nativetest64/bionic-unit-tests/bionic-unit-tests
adb shell /data/nativetest64/bionic-unit-tests-static/bionic-unit-tests-static

Change-Id: I40fb62186ab2fe89ce402a05c36ed2de49d6f701