Revert "[MTE] allocate ring buffer for stack history"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9762789e-f60d-46c3-92ab-3ab0d1f9f1c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1061f94af38030a1dc43d7c08e0c3dd659dea39f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com