Revert "[MTE] allocate ring buffer for stack history"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb79baa9-c475-473f-b109-c78bd41025da). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I90ba9c87144c5733007088fc2377ed273043115a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com