Revert "math_test.cpp: remove now-unnecessary isinff()/isinfl() ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/405cc2b3-7bbd-4a51-b750-81faacc37c30). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I34c6bc0a9acae0b9a26fccbe45ece174cebff1b5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com