Revert "math_test.cpp: remove now-unnecessary isinff()/isinfl() ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c6efce4-3ca9-48cc-bf61-67dbe562c2c0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If1b2c7f5602a4d8d39bc7b9553226421232ec240
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com