Revert "math_test.cpp: remove now-unnecessary isinff()/isinfl() ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4f0ba0c-b98c-4a51-8212-00d43cb1e505). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie160657c949051fb15d8c9bbbbf227977e0dd865
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com