Switch x86 over to the regular socket syscalls.
We haven't needed this since Linux 4.3, which is pretty old now (EOL 2016), and while not particularly harmful, it is one pointless difference between x86 and every other supported architecture.

Where would anyone notice? When writing a seccomp filter. So the "pro" here is "you don't need to know this obscure nonsense in future when writing a seccomp filter". But equally, the "con" is "if you already have an x86-specific seccomp filter that included socketcall(2), you're going to have to update it".

In practice, of the two affected policies in AOSP, only one needs to change. The other was already incorrectly including the regular socket syscalls in addition to socketcall(2) anyway, which for me seems like confirmation that we're better off without this.

Change-Id: Id6e9e72f1181b54ac54cead7fb4c0c477fd8f95a