Revert "math_test.cpp: remove now-unnecessary isinff()/isinfl() ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe1c635a-c16a-4ed3-996f-03034f92aa0d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I210ba638bc8c6c8ab369a8d97121751b138b99cb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com