Improve the <ctype.h> (and <wctype.h>) benchmarks.
From trying to optimize <ctype.h> I've realized that our current
benchmarks -- by reusing the same single input repeatedly -- are
basically just testing the CPU's branch predictor. And modern branch
predictors are pretty good. Subvert this by using randomized data
instead.

The sensible way to do this without duplication is to support benchmark
templates in bionic benchmarks (as regular google-benchmark does), but
that then means we need to remove our reinterpret_cast hack to avoid
pulling in google-benchmark to the benchmark tests. The obvious fix for
that ("don't do that then") that I've gone with here has the unfortunate
side-effect that now our benchmarks are in /data/nativetest rather than
/data/benchmarktest (which in a sense is fine, because I've always felt
like the distinction was arbitrary and unhelpful anyway, but is unusual
compared to everyone else's benchmarks, and a change in our practice too).

Change-Id: I262181ab43e041ce39d9b684071bc945cf996a1c