wctype simplifications.
Remove the unused bogus entry from the array in wctype(), and switch from a switch statement to an array in iswctype(). Oddly (to me) clang is currently completely unrolling the loop in wctype(), which seems pointless, but not recognizing that the switch in iswctype() is just a jump table.

This doesn't change the generated code for wctype() except for removing one unnecessary strcmp() from every invocation, but it changes iswctype() to the optimal code you'd write by hand.

Change-Id: Ib2ac61442c7621288a73f7324d795cf054a60ea5