towctrans() should set errno on failure too.
Technically, POSIX says this is "unspecified" which is presumably why musl/glibc do nothing. But FreeBSD (and thus iOS) set errno to EINVAL, and bionic was already setting errno to EINVAL for wctrans(), so let's be consistent with ourselves and the OS that folks are most likely to be porting code to/from.

(Noticed by inspection.)

Change-Id: I63e252d588ddfb7c6ead58d04c690d7b85f23374