Ignore relocations in the vdso.
glibc and musl both explicitly ignore relocations in the vdso. There
shouldn't be any, but the Ubuntu 24.04 risc-v kernel accidentally
contained relocations, which meant that bionic binaries couldn't start
where glibc/musl ones could.

This shouldn't happen again (because there's now a check as part of the
kernel build), but we might as well behave like the other two libcs that
deal with Linux?

Separately, there shouldn't be any ELF constructors in the vdso, so I've
removed that line too.

Change-Id: I69ca6526e92ac1be0d52ce0d40b2f1091decd6da