Revert "Revert "Workaround app compat issue introduced by global..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b973af50-d203-4c85-b8e2-6955bf2c97f0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I997f0dde8cef71538eb0ee45d7ace1ea5aec7ce7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com