Revert "linker: map large portion of ELF file to read its fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/313f9ab8-434f-47ce-85ea-352d7072cd69). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ied3ebd87d03c75bad4750285b0fba2b545c32b2d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com