Revert "linker: map the entire ELF file if address space is 64bit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/313f9ab8-434f-47ce-85ea-352d7072cd69). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id1146e76f2c58c3d5ebf3bd9fde968668ef1f1b8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com