Revert "linker: map the entire ELF file if address space is 64bit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/313f9ab8-434f-47ce-85ea-352d7072cd69). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id4bf1e4e97c9cf4ccc1cd571d88ec5ca4705c175
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com