Revert "linker: map the entire ELF file if address space is 64bit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a65667b8-d483-4824-aa00-d74bdd060b9e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia6db37c69ac6d86b4abf0243b8eb3d7991a1dde4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com