Revert "linker: map large portion of ELF file to read its fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a65667b8-d483-4824-aa00-d74bdd060b9e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8395f54e39111456f0442121f5743679ed09f0a8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com