Revert "linker: map the entire ELF file if address space is 64bit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5fc28d2a-c8c3-4434-aa16-950d3bd06519). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0e3c8a58e07bf26e55d1c99dd2408455ee17d8f6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com