Revert "linker: map large portion of ELF file to read its fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5fc28d2a-c8c3-4434-aa16-950d3bd06519). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iab93845160c7cd4a94ad51f4fd6625027017199d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com