Revert "linker: map large portion of ELF file to read its fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05264eed-0492-419d-9aaf-0c079d4a2244). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3243a87262abc6db6d81acd0dd5a045015f3a268
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com