Revert "linker: map the entire ELF file if address space is 64bit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05264eed-0492-419d-9aaf-0c079d4a2244). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I12c6d157d62a7fd96f0b9f630e4f81053bb657ba
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com