Revert "linker: map the entire ELF file if address space is 64bit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/07d48818-e8bf-4d67-b2fb-9407e977772d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icad76795dc9490232efe2cf38b76e60f750eb9d8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com