Revert "linker: map large portion of ELF file to read its fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/07d48818-e8bf-4d67-b2fb-9407e977772d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibd39a08caef8b3944b9d218413cb608e30999cfb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com