Revert "linker: map the entire ELF file if address space is 64bit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/db29556f-4b98-4825-aab9-86cd5a68dcdd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I67ba155202f0d0a2ac748ecb52614b365a776ac2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com