Revert "linker: map large portion of ELF file to read its fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/db29556f-4b98-4825-aab9-86cd5a68dcdd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8c6961a74061c111ea27c5635c8dd259630f697a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com