Revert "linker: map large portion of ELF file to read its fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d921e476-feaf-45b7-ba3e-888e5030e9ce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9523c5ea42e9896c2413ec13ee370269606c09dc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com