Revert "linker: map the entire ELF file if address space is 64bit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d921e476-feaf-45b7-ba3e-888e5030e9ce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iee8f2b33536e3b21158c448d54fd08c7de381a1a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com