Revert "linker: map large portion of ELF file to read its fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f70b9b1f-17ab-4aa7-93e9-3996ca4e372b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ice4b66268929d399b00947f9b6bfbe2dc073684a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com