Revert "linker: map large portion of ELF file to read its fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f70b9b1f-17ab-4aa7-93e9-3996ca4e372b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib542cdb0c7c265be9cbce91b72db2b63014a0c13
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com