Revert "linker: map the entire ELF file if address space is 64bit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f70b9b1f-17ab-4aa7-93e9-3996ca4e372b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieee0529c8e13838113fa4213a5bb76bb9e17ea2b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com