Revert "linker: map the entire ELF file if address space is 64bit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/55c83d3f-72bb-40f6-b6c5-394947133ce4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icf834e44f2c80733a9b58c52e26bd4747fe3a119
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com