Revert "linker: map large portion of ELF file to read its fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/55c83d3f-72bb-40f6-b6c5-394947133ce4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I843a05e5924c279fbab44d1544aa329815a99448
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com