Add POSIX qsort_r().
POSIX added this to issue 8, based on https://www.austingroupbugs.net/view.php?id=900.

This is an incompatibility with iOS, which is one reason why I hadn't
added this previously, but FreeBSD (upstream for both Android and iOS in
this case) already moved off their different qsort_r(), so it's possible
Apple will, and in the meantime, at least I now have the defense of
"POSIX says...".

Change-Id: Iab048a3a931010c3276a88d25eb56b38d4d608b1