Revert "linker: map large portion of ELF file to read its fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bdde5f6d-799d-4327-b5a4-555d81e1d600). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic684914d14b3ff22d2c4e9af5253605a6d26c647
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com