Revert "linker: map the entire ELF file if address space is 64bit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bdde5f6d-799d-4327-b5a4-555d81e1d600). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8579aae245154e4ab74900b55cc1992b0d310b42
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com