Revert "linker: map large portion of ELF file to read its fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bdde5f6d-799d-4327-b5a4-555d81e1d600). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I78560e1cf213af15aae48cb028a9de559a22cf87
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com