async_safe_test.cpp: `__ANDROID_API_LEVEL__` isn't a thing.
`__ANDROID_API__` is the actual name for this. There's another reference to this in the _code_ too, but fixing that currently breaks the build because the vendor ABI dumps don't yet include the new function we're guarding the use of.

In the meantime, at least make sure we're building and running the tests for libc.

Change-Id: Ib7276948812cb660ce765da61ab1ea97fa282185