Revert "Revert "Workaround app compat issue introduced by global..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/03c2dc3f-9a99-4cdf-9da5-b189382c4545). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 354495519
Change-Id: Idd428ef4efec4e0b6ee9da48cf002327318c77b3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com