Revert "Revert "Workaround app compat issue introduced by global..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd6f4e87-c73c-419e-ba84-7de7a217d03d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 354495519
Change-Id: If7e4773130c0ba863179cbf09e0677025356c875
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com