Revert "Revert "linker: map large portion of ELF file to read it..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/123396d5-2970-467e-9204-b78db79f599a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6f53d9ca21f2f56dfc860b81b56651548158837f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com