Revert "libc.map.txt: remove the two riscv64 special cases."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/123396d5-2970-467e-9204-b78db79f599a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1c47a4100c83a03407c776e71621eb7976a62cb3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com