Bionic: libc: update bpf.h raw_tracepoint
Upstream kernel commit [1] adds a new field to the raw_tracepoint struct
in bpf.h. In updating //external/libbpf to the latest version, the build
fails due to the missing field.

Adding this field to userspace code should not effect the older kernel
sides, as the bpf syscalls are designed to accept structures larger than
they know about.

[1] https://lore.kernel.org/all/20240319233852.1977493-4-andrii@kernel.org/

Test: m
Change-Id: I5ac0ff7f4ecfeb7e130c337b69cdedf6d68f3441
Signed-off-by: Neill Kapron <nkapron@google.com>