Revert "Linker: clear error for incorrect page size load."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/977ae95a-d970-4e70-8db3-15cc33faa2e8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icdb5e791105ca84c3cdfb74f33eec07dc7e6e85d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com