Revert "linker: map large portion of ELF file to read its fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/675c761a-a8f9-4a82-ab01-3ce1844d0ce8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1ad84b6514dbbb2b99f49a113065707049c7c6df
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com