Revert "linker: map the entire ELF file if address space is 64bit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/675c761a-a8f9-4a82-ab01-3ce1844d0ce8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I35ca5f9e0e69e61c6f38ea546fe82c7f06effdf8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com