Revert "Add mseal(2) wrapper."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0bb03238-a1ed-4048-989b-8da0ca5b418c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia96be6631b5c145a10e0edc2a789fc090d115a89
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com