Revert "Stop using versioner."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/07df58ca-7aae-4ed2-9734-c0fd2ebbc6e0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361440272
Bug: 361442729
Change-Id: I29934875339baecceac08136b337ddce25698866
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com