Revert "Stop using versioner."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8f799f2b-2d33-4999-8ed2-86b49b2710b9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361443018
Change-Id: Ie5a359875383cf76561e95664d0c0183d7fde4a0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com