Revert "Sync with upstream FreeBSD libm."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ac0945b9-1fad-49c8-af28-1337608765a3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie59794927271b3f28317b35beced15646da47156
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com